Refactor repeated test logic into methods
