+++ tests/test_twitter_bot/utils.py
@@ -1,0 +1,11 @@
+from settings import TWITTER_USER_USERNAME
+from twitter_api_client import TwitterAPIClient
+
+
+def delete_created_tweets():
+    """Delete created Tweets from Twitter."""
+    twitter_api = TwitterAPIClient()
+    for tweet in twitter_api.get_recent_tweets(TWITTER_USER_USERNAME):
+        tweet_id = tweet["id_str"]
+        twitter_api.delete_tweet(tweet_id)
+        assert not twitter_api.tweet_exists(tweet_id)
